GraphService: refactor
